# api/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "API for Objective Zero"
  version: "1.0.0"
  title: "Objective Zero API"
  contact:
    email: "you@your-company.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "107.170.82.21:3000"
basePath: "/mobile"
...
schemes:
- "http"
paths:
...
  /updateambassador:
    post:
      summary: "Update the Ambassador profile"
      description: "Ambassador can update his/her profile"
      operationId: "updateambassador"
      produces:
      - "application/json"
      parameters:
        - name: access_token
          in: header
          description: Access Token
          required: true
          type: string
          format: string
        - in: body
          name: Update Ambassador profile
          description: Profile Update.
          schema:
            type: object
            properties:
              name:
                type: string 
              age:
                type: string
              gender:
                type: string
              service_member:
                type: string
              service:
                type: string
              veteran:
                type: string
              mos:
                type: string
              component:
                type: string             
      responses:
        200:
          description: "search results matching criteria"
        400:
          description: "bad input parameter"
      x-swagger-router-controller: "add_ambassador"
  /getambassadorprofile:
    post:
      summary: "get the Ambassador Profile"
      description: "Ambassador can get their profile"
      operationId: "getambassadorprofile"
      produces:
      - "application/json"
      parameters:
        - name: access_token
          in: header
          description: Access Token
          required: true
          type: string
          format: string            
      responses:
        200:
          description: "search results matching criteria"
        400:
          description: "bad input parameter"
      x-swagger-router-controller: "add_ambassador"
...